Extract query value helpers
